\name{runfptest}
\alias{runfptest}
\encoding{UTF-8}
\title{Run floating point tests in R.}
\description{Simple routine to run tests from within an R session. }
\usage{
   runfptest(control=list())
}
\arguments{
 \item{control}{
    An optional list of control settings.
 }
}

\details{
  There are currently no details. See vignettes.
}

\value{
   Returns a character type message. To be determined.
}
\keyword{floating point}

